+++ flatlaf-core/src/main/java/com/formdev/flatlaf/ui/FlatTextFieldUI.java
@@ -252,7 +252,7 @@ public class FlatTextFieldUI
 	@Override
 	protected void paintSafely( Graphics g ) {
 		paintBackground( g, getComponent(), isIntelliJTheme, focusedBackground );
-		paintPlaceholder( g, getComponent(), placeholderForeground );
+		paintPlaceholder( g );
 
 		super.paintSafely( HiDPIUtils.createGraphicsTextYCorrection( (Graphics2D) g ) );
 	}
@@ -308,7 +308,9 @@ public class FlatTextFieldUI
 		return background;
 	}
 
-	static void paintPlaceholder( Graphics g, JTextComponent c, Color placeholderForeground ) {
+	protected void paintPlaceholder( Graphics g ) {
+		JTextComponent c = getComponent();
+
 		// check whether text component is empty
 		if( c.getDocument().getLength() > 0 )
 			return;
@@ -323,16 +325,14 @@ public class FlatTextFieldUI
 			return;
 
 		// compute placeholder location
-		Insets insets = c.getInsets();
+		Rectangle r = getVisibleEditorRect();
 		FontMetrics fm = c.getFontMetrics( c.getFont() );
-		int x = insets.left;
-		int y = insets.top + fm.getAscent() + ((c.getHeight() - insets.top - insets.bottom - fm.getHeight()) / 2);
+		int y = r.y + fm.getAscent() + ((r.height - fm.getHeight()) / 2);
 
 		// paint placeholder
 		g.setColor( placeholderForeground );
-		String clippedPlaceholder = JavaCompatibility.getClippedString( jc, fm,
-				(String) placeholder, c.getWidth() - insets.left - insets.right );
-		FlatUIUtils.drawString( c, g, clippedPlaceholder, x, y );
+		String clippedPlaceholder = JavaCompatibility.getClippedString( c, fm, (String) placeholder, r.width );
+		FlatUIUtils.drawString( c, g, clippedPlaceholder, r.x, y );
 	}
 
 	@Override
@@ -384,5 +384,14 @@ public class FlatTextFieldUI
 	protected Insets getPadding() {
 		Object padding = getComponent().getClientProperty( FlatClientProperties.TEXT_FIELD_PADDING );
 		return (padding instanceof Insets) ? UIScale.scale( (Insets) padding ) : null;
+	}
+
+	/**
+	 * @since 1.4
+	 */
+	protected void scrollCaretToVisible() {
+		Caret caret = getComponent().getCaret();
+		if( caret instanceof FlatCaret )
+			((FlatCaret)caret).scrollCaretToVisible();
 	}
 }