+++ flatlaf-core/src/main/java/com/formdev/flatlaf/ui/FlatToolBarUI.java
@@ -25,6 +25,7 @@ import java.beans.PropertyChangeListener;
 import java.util.Map;
 import javax.swing.AbstractButton;
 import javax.swing.JComponent;
+import javax.swing.UIManager;
 import javax.swing.border.Border;
 import javax.swing.plaf.ComponentUI;
 import javax.swing.plaf.basic.BasicToolBarUI;
@@ -45,6 +46,10 @@ import com.formdev.flatlaf.ui.FlatStyleSupport.Styleable;
  * @uiDefault ToolBar.floatingForeground				Color
  * @uiDefault ToolBar.isRollover						boolean
  *
+ * <!-- FlatToolBarUI -->
+ *
+ * @uiDefault ToolBar.focusableButtons					boolean
+ *
  * <!-- FlatToolBarBorder -->
  *
  * @uiDefault ToolBar.borderMargins				Insets
@@ -55,6 +60,9 @@ import com.formdev.flatlaf.ui.FlatStyleSupport.Styleable;
 public class FlatToolBarUI
 	extends BasicToolBarUI
 {
+	/** @since 1.4 */
+	protected boolean focusableButtons;
+
 	// for FlatToolBarBorder
 	@Styleable protected Insets borderMargins;
 	@Styleable protected Color gripColor;
@@ -69,6 +77,10 @@ public class FlatToolBarUI
 	public void installUI( JComponent c ) {
 		super.installUI( c );
 
+		// disable focusable state of buttons (when switching from another Laf)
+		if( !focusableButtons )
+			setButtonsFocusable( false );
+
 		applyStyle( FlatStyleSupport.getStyle( c ) );
 	}
 
@@ -76,7 +88,18 @@ public class FlatToolBarUI
 	public void uninstallUI( JComponent c ) {
 		super.uninstallUI( c );
 
+		// re-enable focusable state of buttons (when switching to another Laf)
+		if( !focusableButtons )
+			setButtonsFocusable( true );
+
 		oldStyleValues = null;
+	}
+
+	@Override
+	protected void installDefaults() {
+		super.installDefaults();
+
+		focusableButtons = UIManager.getBoolean( "ToolBar.focusableButtons" );
 	}
 
 	@Override
@@ -86,18 +109,22 @@ public class FlatToolBarUI
 			public void componentAdded( ContainerEvent e ) {
 				super.componentAdded( e );
 
-				Component c = e.getChild();
-				if( c instanceof AbstractButton )
-					c.setFocusable( false );
+				if( !focusableButtons ) {
+					Component c = e.getChild();
+					if( c instanceof AbstractButton )
+						c.setFocusable( false );
+				}
 			}
 
 			@Override
 			public void componentRemoved( ContainerEvent e ) {
 				super.componentRemoved( e );
 
-				Component c = e.getChild();
-				if( c instanceof AbstractButton )
-					c.setFocusable( true );
+				if( !focusableButtons ) {
+					Component c = e.getChild();
+					if( c instanceof AbstractButton )
+						c.setFocusable( true );
+				}
 			}
 		};
 	}
@@ -119,6 +146,16 @@ public class FlatToolBarUI
 	 */
 	protected Object applyStyleProperty( String key, Object value ) {
 		return FlatStyleSupport.applyToAnnotatedObject( this, key, value );
+	}
+
+	/**
+	 * @since 1.4
+	 */
+	protected void setButtonsFocusable( boolean focusable ) {
+		for( Component c : toolBar.getComponents() ) {
+			if( c instanceof AbstractButton )
+				c.setFocusable( focusable );
+		}
 	}
 
 	// disable rollover border